Refactor: API 경로 수정 #71
